+++ app/models/schema/response.py
@@ -2,15 +2,10 @@ from pydantic import BaseModel
 from typing import Optional
 from enum import Enum
 
+
 class ApiResponseStatus(Enum):
     success = "success"
     failed = "failed"
-
-
-class CommonPathParams:
-    def __init__(self, season: int, league_id: int) -> None:
-        self.season = season
-        self.league_id = league_id
 
 
 class ApiResponse(BaseModel):
